Rifat

+++ app/Http/Controllers/PersonalInfoController.php
@@ -244,7 +244,8 @@ class PersonalInfoController extends Controller
         $employee->save();
 
         if($r->hasFile('image')){
-            if(file_exists(public_path('candidateImages/'.$employee->image))){
+            if($employee->image != '' && file_exists(public_path('candidateImages/'.$employee->image)) && file_exists(public_path('candidateImages/thumb/'.$employee->image))){
+                unlink(public_path('candidateImages/thumb/'.$employee->image));
                 unlink(public_path('candidateImages/'.$employee->image));
             }
             $img = $r->file('image');
@@ -259,7 +260,8 @@ class PersonalInfoController extends Controller
         }
 
         if($r->hasFile('sign')){
-            if(file_exists(public_path('candidateSigns/'.$employee->sign))){
+            if($employee->sign != '' && file_exists(public_path('candidateSigns/'.$employee->sign)) && file_exists(public_path('candidateSigns/thumb/'.$employee->sign))){
+                unlink(public_path('candidateSigns/thumb/'.$employee->sign));
                 unlink(public_path('candidateSigns/'.$employee->sign));
             }
             $sign = $r->file('sign');
